chore(origin-ui): split user tests
