+++ packages/apps/origin-ui/cypress.env.json.example
@@ -5,7 +5,7 @@
         "firstName": "Test",
         "lastName": "User",
         "telephone": "+1234567890",
-        "title: "Mr"
+        "title": "Mr"
     },
     "adminUser": {
         "email": "admin@mailinator.com",
